+++ services/autoscaling/src/simcore_service_autoscaling/modules/cluster_scaling/_buffer_machines_pool_core.py
@@ -52,7 +52,7 @@ from ...utils.buffer_machines_pool_core import (
 from ..ec2 import get_ec2_client
 from ..instrumentation import get_instrumentation, has_instrumentation
 from ..ssm import get_ssm_client
-from .auto_scaling_mode_base import AutoscalingProvider
+from ._provider_protocol import AutoscalingProvider
 
 _logger = logging.getLogger(__name__)
